+++ src/Streamly/Internal/Data/Stream/Channel/Dispatcher.hs
@@ -9,14 +9,16 @@
 --
 module Streamly.Internal.Data.Stream.Channel.Dispatcher
     (
-    -- ** Dispatching
-      pushWorker
+    -- *** Worker Dispatching
+    -- | Low level functions used to build readOutputQ and postProcess
+    -- functions.
+      forkWorker
     , dispatchWorker
     , dispatchWorkerPaced
-    , sendWorkerWait
-    , startChannel
+    , dispatchAllWait
     , sendWorkerDelay
     , sendWorkerDelayPaced
+    , startChannel -- XXX bootstrap?
     )
 where
 
@@ -42,14 +44,22 @@ import Streamly.Internal.Data.Stream.Channel.Type
 -- Dispatching workers
 -------------------------------------------------------------------------------
 
-{-# NOINLINE pushWorker #-}
-pushWorker :: MonadRunInIO m => Count -> Channel m a -> m ()
-pushWorker yieldMax sv = do
+-- | Low level API to create a worker. Forks a thread which executes the
+-- 'workLoop' of the channel.
+{-# NOINLINE forkWorker #-}
+forkWorker :: MonadRunInIO m =>
+       Count -- ^ max yield limit for the worker
+    -> Channel m a
+    -> m ()
+forkWorker yieldMax sv = do
     liftIO $ atomicModifyIORefCAS_ (workerCount sv) $ \n -> n + 1
     when (svarInspectMode sv)
         $ recordMaxWorkers (workerCount sv) (svarStats sv)
     -- This allocation matters when significant number of workers are being
     -- sent. We allocate it only when needed.
+    --
+    -- XXX WorkerInfo is required for maxYields to work even if rate control is
+    -- not enabled.
     winfo <-
         case yieldRateInfo sv of
             Nothing -> return Nothing
@@ -130,8 +140,17 @@ checkMaxBuffer active sv = do
             (_, n) <- liftIO $ readIORef (outputQueue sv)
             return $ fromIntegral lim > n + active
 
+-- | Higher level API to dispatch a worker, it uses 'forkWorker' to create a
+-- worker. Dispatches a worker only if:
+--
+-- * the channel has work to do
+-- * max thread count is not reached
+-- * max buffer limit is not reached
+--
 dispatchWorker :: MonadRunInIO m =>
-    Count -> Channel m a -> m Bool
+       Count -- ^ max yield limit for the worker
+    -> Channel m a
+    -> m Bool -- ^ can disptach more workers
 dispatchWorker yieldCount sv = do
     -- XXX in case of Ahead streams we should not send more than one worker
     -- when the work queue is done but heap is not done.
@@ -157,13 +176,13 @@ dispatchWorker yieldCount sv = do
             then do
                 r1 <- checkMaxBuffer active sv
                 if r1
-                then pushWorker yieldCount sv >> return True
+                then forkWorker yieldCount sv >> return True
                 else return False
             else return False
         else do
             when (active <= 0) $ do
                 r <- liftIO $ isWorkDone sv
-                when (not r) $ pushWorker 0 sv
+                when (not r) $ forkWorker 0 sv
             return False
     else return False
 
@@ -172,11 +191,10 @@ dispatchWorker yieldCount sv = do
 -- where we keep dispatching and they keep returning. So we must have exactly
 -- the same logic for not dispatching and for returning.
 
--- | Dispatcher with rate control. The number of workers to be dispatched are
--- decided based on the target rate.
+-- | Like 'dispatchWorker' but with rate control. The number of workers to be
+-- dispatched are decided based on the target rate. Uses 'dispatchWorker' to
+-- actually dispatch when required.
 --
--- This is called when reading the output queue of the channel and after all
--- the items read in one batch are all processed.
 dispatchWorkerPaced :: MonadRunInIO m =>
        Channel m a
     -> m Bool -- ^ True means can dispatch more
@@ -289,15 +307,22 @@ dispatchWorkerPaced sv = do
             then dispatchN (n - 1)
             else return False
 
-{-# NOINLINE sendWorkerWait #-}
-sendWorkerWait
+-- | Dispatches as many workers as it can until output is seen in the event
+-- queue of the channel. If the dispatcher function returns 'False' then no
+-- more dispatches can be done. If no more dispatches are possible blocks until
+-- output arrives in the event queue.
+--
+-- When this function returns we are sure that there is some output available.
+--
+{-# NOINLINE dispatchAllWait #-}
+dispatchAllWait
     :: MonadIO m
-    => Bool
-    -> (Channel m a -> IO ())
-    -> (Channel m a -> m Bool)
+    => Bool -- ^ 'eager' option is on
+    -> (Channel m a -> IO ()) -- ^ delay function
+    -> (Channel m a -> m Bool) -- ^ dispatcher function
     -> Channel m a
     -> m ()
-sendWorkerWait eagerEval delay dispatch sv = go
+dispatchAllWait eagerEval delay dispatch sv = go
 
     where
 
@@ -375,7 +400,7 @@ sendWorkerWait eagerEval delay dispatch sv = go
                     $ withDiagMVar
                         (svarInspectMode sv)
                         (dumpChannel sv)
-                        "sendWorkerWait: nothing to do"
+                        "dispatchAllWait: nothing to do"
                     $ takeMVar (outputDoorBell sv)
                 (_, len) <- liftIO $ readIORef (outputQueue sv)
                 if len <= 0
@@ -392,15 +417,17 @@ startChannel :: MonadRunInIO m =>
     Channel m a -> m ()
 startChannel chan = do
     case yieldRateInfo chan of
-        Nothing -> pushWorker 0 chan
+        Nothing -> forkWorker 0 chan
         Just yinfo  ->
             if svarLatencyTarget yinfo == maxBound
             then liftIO $ threadDelay maxBound
-            else pushWorker 1 chan
+            else forkWorker 1 chan
 
+-- | Noop as of now.
 sendWorkerDelayPaced :: Channel m a -> IO ()
 sendWorkerDelayPaced _ = return ()
 
+-- | Noop as of now.
 sendWorkerDelay :: Channel m a -> IO ()
 sendWorkerDelay _sv =
     -- XXX we need a better way to handle this than hardcoded delays. The