+++ docs/User/Project/Changelog.md
@@ -2,6 +2,10 @@
 
 <!-- See rendered changelog at https://streamly.composewell.com -->
 
+## unreleased
+
+* Fix rate control for ordered streams.
+
 ## 0.10.1 (Jan 2024)
 
 * Fix TH macros in `Streamly.Data.Stream.MkType` for GHC 9.6 and above.